+++ project/PhongMaterial.hpp
@@ -4,12 +4,12 @@
 
 class PhongMaterial {
 public:
-    PhongMaterial(const glm::vec3& kd, const glm::vec3& ks, double shininess);
+    PhongMaterial(const glm::dvec3& kd, const glm::dvec3& ks, double shininess);
 
     virtual ~PhongMaterial();
 
-    const glm::vec3 m_kd;
-    const glm::vec3 m_ks;
+    const glm::dvec3 m_kd;
+    const glm::dvec3 m_ks;
 
     const double m_shininess;
 };
